Add balance validity service (#129)
* Rename block synchronizer

* Add types of witness

* Modify ValidityWitness

* wip

* Sync deposit tree

* Sync account tree

* small fix

* Implement validity function (wip)

* Fix Merkle tree

* Fix by linter

* Remove unnecessary file

* Connect block validity prover

* small fix

* Fix /transaction API

* Fix ticker

* Add private balance validity prover

* Modify private balance prover

* Modify private balance validity prover

* wip

* Revert "wip"

This reverts commit ec0503dfe7536efac870a47cfafe85f14072a177.

* Add database definition

* Add historical Merkle tree

* Fix Merkle tree

* wip

* wip

* wip

* Fix block validity prover

* Add deposits table

* wip

* wip

* wip

* Fix deposit_index column

* wip

* Fix prover

* Test using mock wallet and sample data (#126)

* Create sample data

* Pass e2e test

* Add request ID in ProofResponse

* wip

* Integrate withdrawal prover

* Work Gnark server

* Fix submitWithdrawalProof

* wip

* Add SyncLocally

* wip

* Separate client side code

* Remove SyncBlockProverWithBlockNumber from client code

* wip

* FIx block update

* Modify MockWallet

* wip

* Pass deposit proof

* wip

* Fix nonce type

* Remove BlockNumberByDepositIndex

* Fix txDetails

+++ internal/balance_prover_service/types.go
@@ -126,7 +126,7 @@ func (input *ValidityPublicInputsInput) FromValidityPublicInputs(value *block_va
 
 type TxInput struct {
 	TransferTreeRoot poseidonHashOut `json:"transferTreeRoot"`
-	Nonce            uint64          `json:"nonce"`
+	Nonce            uint32          `json:"nonce"`
 }
 
 type TxWitnessInput struct {
@@ -296,7 +296,7 @@ func NewSpentValue(
 	}
 
 	insufficientFlags = backup_balance.InsufficientFlags(insufficientFlags)
-	isValid := txNonce == uint64(prevPrivateState.Nonce)
+	isValid := uint32(txNonce) == prevPrivateState.Nonce
 	newPrivateState := PrivateState{
 		AssetTreeRoot:     assetTreeRoot,
 		Nonce:             prevPrivateState.Nonce + 1,
@@ -313,7 +313,7 @@ func NewSpentValue(
 	transferRoot, _, _ := transferTree.GetCurrentRootCountAndSiblings()
 	tx := intMaxTypes.Tx{
 		TransferTreeRoot: &transferRoot,
-		Nonce:            txNonce,
+		Nonce:            uint32(txNonce),
 	}
 
 	return &SpentValue{
@@ -346,7 +346,7 @@ func (w *SendWitness) GetNextLastTx() (*SendWitnessResult, error) {
 		w.NewPrivateStateSalt,
 		w.Transfers,
 		w.AssetMerkleProofs,
-		w.TxWitness.Tx.Nonce,
+		uint64(w.TxWitness.Tx.Nonce),
 	)
 	if err != nil {
 		return nil, err
